ref(conversations): Move shared AI modules back to insights
The previous two commits moved several modules out of insights/ into
explore/conversations/ that are still consumed by the insights/pages/agents/
pages (overview, onboarding, drawer, modelCostWidget). Leaving those
modules under conversations/ forces insights to cross into the explore
tree for its own working files, which is the wrong direction.

Revert those specific moves; keep in conversations only the modules that
have zero insights callers. Callers outside insights (conversations
itself, performance/newTraceDetails/traceDrawer, gettingStartedDocs)
continue to use the insights paths.

Moved back to insights:
  explore/conversations/utils/getDefaultSelectedNode.ts
    → insights/pages/agents/utils/getDefaultSelectedNode.tsx
  explore/conversations/utils/formatLLMCosts.ts
    → insights/pages/agents/utils/formatLLMCosts.tsx
  explore/conversations/utils/onboardingUtils.tsx
    → insights/pages/onboardingUtils.tsx
  explore/conversations/utils/agentIntegrations.ts
    → insights/pages/agents/utils/agentIntegrations.tsx
  explore/conversations/components/llmOnboardingInstructions.tsx (+ spec)
    → insights/pages/agents/llmOnboardingInstructions.tsx (+ spec)
  explore/conversations/components/llmCosts.tsx
    → insights/pages/agents/components/llmCosts.tsx
  explore/conversations/components/aiSpanList.tsx
    → insights/pages/agents/components/aiSpanList.tsx
  explore/conversations/components/agentSelector.tsx
    → insights/common/components/agentSelector.tsx

Kept in conversations (no insights callers):
  explore/conversations/utils/features.ts
  explore/conversations/utils/timeBounds.ts (+ spec)

Both are consumed only by conversations and
performance/newTraceDetails/traceDrawer/tabs/; neither has any insights
importer.
