+++ static/app/views/explore/conversations/hooks/useConversationSelection.tsx
@@ -2,7 +2,7 @@ import {useCallback, useEffect, useMemo} from 'react';
 
 import {useFocusedToolSpan} from 'sentry/views/explore/conversations/hooks/useFocusedToolSpan';
 import {extractMessagesFromNodes} from 'sentry/views/explore/conversations/utils/conversationMessages';
-import {getDefaultSelectedNode} from 'sentry/views/explore/conversations/utils/getDefaultSelectedNode';
+import {getDefaultSelectedNode} from 'sentry/views/insights/pages/agents/utils/getDefaultSelectedNode';
 import type {AITraceSpanNode} from 'sentry/views/insights/pages/agents/utils/types';
 
 interface UseConversationSelectionOptions {
