+++ static/app/views/insights/pages/agents/components/aiSpanList.tsx
@@ -11,7 +11,7 @@ import {IconChat, IconChevron, IconCode, IconFire, IconFix} from 'sentry/icons';
 import {IconBot} from 'sentry/icons/iconBot';
 import {t} from 'sentry/locale';
 import {getDuration} from 'sentry/utils/duration/getDuration';
-import {LLMCosts} from 'sentry/views/explore/conversations/components/llmCosts';
+import {LLMCosts} from 'sentry/views/insights/pages/agents/components/llmCosts';
 import {
   getFirstToolInputValue,
   getGenAiOpType,
