+++ static/app/views/insights/pages/agents/llmOnboardingInstructions.tsx
