+++ static/app/views/insights/pages/agents/onboarding.tsx
@@ -48,17 +48,20 @@ import {useLocation} from 'sentry/utils/useLocation';
 import {useNavigate} from 'sentry/utils/useNavigate';
 import {useOrganization} from 'sentry/utils/useOrganization';
 import {useProjects} from 'sentry/utils/useProjects';
+import {useSpans} from 'sentry/views/insights/common/queries/useDiscover';
 import {
   CopyLLMPromptButton,
   LLM_ONBOARDING_COPY_MARKDOWN,
-} from 'sentry/views/explore/conversations/components/llmOnboardingInstructions';
+} from 'sentry/views/insights/pages/agents/llmOnboardingInstructions';
 import {
   AGENT_INTEGRATION_ICONS,
   AGENT_INTEGRATION_LABELS,
   DENO_AGENT_INTEGRATIONS,
   NODE_AGENT_INTEGRATIONS,
   PYTHON_AGENT_INTEGRATIONS,
-} from 'sentry/views/explore/conversations/utils/agentIntegrations';
+} from 'sentry/views/insights/pages/agents/utils/agentIntegrations';
+import {getHasAiSpansFilter} from 'sentry/views/insights/pages/agents/utils/query';
+import {Referrer} from 'sentry/views/insights/pages/agents/utils/referrers';
 import {
   BulletList,
   HeaderText,
@@ -66,10 +69,7 @@ import {
   PulsingIndicator,
   SubTitle,
   useOnboardingProject,
-} from 'sentry/views/explore/conversations/utils/onboardingUtils';
-import {useSpans} from 'sentry/views/insights/common/queries/useDiscover';
-import {getHasAiSpansFilter} from 'sentry/views/insights/pages/agents/utils/query';
-import {Referrer} from 'sentry/views/insights/pages/agents/utils/referrers';
+} from 'sentry/views/insights/pages/onboardingUtils';
 
 function useAiSpanWaiter(project: Project) {
   const {selection} = usePageFilters();
