+++ static/app/views/insights/pages/agents/utils/formatLLMCosts.tsx
