+++ static/app/views/insights/pages/agents/utils/getDefaultSelectedNode.tsx
