+++ static/app/views/insights/pages/onboardingUtils.tsx
